fixed messageid type, added default data for models, change a bit user load to immediatly save user to db, aded input for account data (now i need to rewrite db singleton into nothing so everything could start working)

+++ HealthBot/Migrations/20240115062039_Last_Message_Id.cs
@@ -13,7 +13,7 @@ namespace HealthBot.Migrations
             migrationBuilder.AddColumn<int>(
                 name: "messageid",
                 table: "users",
-                type: "bigint",
+                type: "integer",
                 nullable: false,
                 defaultValue: 0
             );
